move dictionary files to dictionary folder
