+++ .eslintrc.json
@@ -713,7 +713,7 @@
                 "ext/js/general/regex-util.js",
                 "ext/js/general/text-source-map.js",
                 "ext/js/language/deinflector.js",
-                "ext/js/language/dictionary-database.js",
+                "ext/js/dictionary/dictionary-database.js",
                 "ext/js/language/sandbox/dictionary-data-util.js",
                 "ext/js/language/sandbox/japanese-util.js",
                 "ext/js/language/translator.js",
@@ -741,10 +741,10 @@
                 "ext/js/data/database.js",
                 "ext/js/data/json-schema.js",
                 "ext/js/general/cache-map.js",
-                "ext/js/language/dictionary-database.js",
-                "ext/js/language/dictionary-importer.js",
-                "ext/js/language/dictionary-worker-handler.js",
-                "ext/js/language/dictionary-worker-media-loader.js",
+                "ext/js/dictionary/dictionary-database.js",
+                "ext/js/dictionary/dictionary-importer.js",
+                "ext/js/dictionary/dictionary-worker-handler.js",
+                "ext/js/dictionary/dictionary-worker-media-loader.js",
                 "ext/js/media/media-util.js"
             ],
             "env": {
