+++ dev/jsconfig.json
@@ -64,8 +64,8 @@
         "../ext/js/general/regex-util.js",
         "../ext/js/general/text-source-map.js",
         "../ext/js/language/deinflector.js",
-        "../ext/js/language/dictionary-importer.js",
-        "../ext/js/language/dictionary-database.js",
+        "../ext/js/dictionary/dictionary-importer.js",
+        "../ext/js/dictionary/dictionary-database.js",
         "../ext/js/language/sandbox/dictionary-data-util.js",
         "../ext/js/language/sandbox/japanese-util.js",
         "../ext/js/language/translator.js",
