+++ ext/js/background/offscreen.js
@@ -21,7 +21,7 @@ import {ClipboardReader} from '../comm/clipboard-reader.js';
 import {invokeMessageHandler} from '../core.js';
 import {createApiMap, getApiMapHandler} from '../core/api-map.js';
 import {ArrayBufferUtil} from '../data/sandbox/array-buffer-util.js';
-import {DictionaryDatabase} from '../language/dictionary-database.js';
+import {DictionaryDatabase} from '../dictionary/dictionary-database.js';
 import {JapaneseUtil} from '../language/sandbox/japanese-util.js';
 import {Translator} from '../language/translator.js';
