+++ test/fixtures/translator-test.js
@@ -24,8 +24,8 @@ import {expect, vi} from 'vitest';
 import {parseJson} from '../../dev/json.js';
 import {createDictionaryArchive} from '../../dev/util.js';
 import {AnkiNoteDataCreator} from '../../ext/js/data/sandbox/anki-note-data-creator.js';
-import {DictionaryDatabase} from '../../ext/js/language/dictionary-database.js';
-import {DictionaryImporter} from '../../ext/js/language/dictionary-importer.js';
+import {DictionaryDatabase} from '../../ext/js/dictionary/dictionary-database.js';
+import {DictionaryImporter} from '../../ext/js/dictionary/dictionary-importer.js';
 import {JapaneseUtil} from '../../ext/js/language/sandbox/japanese-util.js';
 import {Translator} from '../../ext/js/language/translator.js';
 import {chrome, fetch} from '../mocks/common.js';
