+++ types/ext/translator.d.ts
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-import type {DictionaryDatabase} from '../../ext/js/language/dictionary-database';
+import type {DictionaryDatabase} from '../../ext/js/dictionary/dictionary-database';
 import type {JapaneseUtil} from '../../ext/js/language/sandbox/japanese-util';
 import type * as Dictionary from './dictionary';
 import type * as DictionaryDatabaseTypes from './dictionary-database';
